Better handling on collision enter and on collision exit. But with errors. Set is in collision with player on objects.

+++ YasEngine/Button.cpp
@@ -29,7 +29,7 @@ Button::Button(const ButtonId& buttonId, std::string text, const Vector4D<Uint8>
 {
     this->buttonId = buttonId;
     this->text.assign(text);
-    this->buttonTextWidth = ScreenWriter::FONT_WIDTH * text.size();
+    this->buttonTextWidth = ScreenWriter::FONT_WIDTH * static_cast<int>(text.size());
     this->color = color;
     numberOfVertices = 4;
 }
